Add matching from identifier to place category
Since we have an index mapping the identifiers to places,
we want to use that for categorisation purposes.

This one is a funny one. The index_converter takes the index provided
by the archives and provides us with "identifier --> place".
On the other hand, using the usual manual mapping way,
we have a 'place --> category' mapping along with the other mappings
(ie in the mapper data structure, which is used in post-processing).
And for various reasons, we want to match all that in the post_processing!

Enter match_identifier_to_categories, which is called the usual way by
post_process_collection. It thus drags along in the mapper the second mapping.
We retrieve from the index module the first mapping.
And we just have to glue everything together.